+++ app.go
@@ -1,12 +1,10 @@
 package main 
 
 import (
-	"log-collection-api/middleware"
+    "log-collection-api/pkg/routes"
 	"log"
 	"net/http"
 
-	jwtmiddleware "github.com/auth0/go-jwt-middleware/v2"
-	"github.com/auth0/go-jwt-middleware/v2/validator"
     "github.com/rs/cors"
     "github.com/unrolled/secure"
 )
@@ -23,44 +21,8 @@ type App struct {
 }
 
 func (app *App) StartServer() {
-	router := http.NewServeMux()
-	// This route is always accessible.
-	router.Handle("/api/messages/public", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-        log.Printf("hello from %s",r)
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusOK)
-		w.Write([]byte(`{"message":"Hello from a public endpoint! You don't need to be authenticated to see this."}`))
-	}))
 
-	// This route is only accessible if the user has a valid access_token.
-	router.Handle("/api/messages/private", middleware.EnsureValidToken()(
-		http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			w.Header().Set("Content-Type", "application/json")
-			w.WriteHeader(http.StatusOK)
-			w.Write([]byte(`{"message":"Hello from a private endpoint! You need to be authenticated to see this."}`))
-		}),
-	))
-
-	// This route is only accessible if the user has a
-	// valid access_token with the read:messages scope.
-	router.Handle("/api/messages/private-scoped", middleware.EnsureValidToken()(
-		http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			w.Header().Set("Content-Type", "application/json")
-
-			token := r.Context().Value(jwtmiddleware.ContextKey{}).(*validator.ValidatedClaims)
-
-			claims := token.CustomClaims.(*middleware.CustomClaims)
-			if !claims.HasScope("read:messages") {
-				w.WriteHeader(http.StatusForbidden)
-				w.Write([]byte(`{"message":"Insufficient scope."}`))
-				return
-			}
-
-			w.WriteHeader(http.StatusOK)
-			w.Write([]byte(`{"message":"Hello from a private endpoint! You need to be authenticated to see this."}`))
-		}),
-	))
-
+    router := routes.Router()
     corsMiddleware := cors.New(app.Config.CorsOptions)
     routerWithCORS := corsMiddleware.Handler(router)
